cleanup: reduce bloatedness of seslib/__init__.py
Before:

$ wc --lines seslib/__init__.py
2195 seslib/__init__.py

After:

$ wc --lines seslib/__init__.py
1536 seslib/__init__.py
